+++ spec/lib/fixtures/dot_slash_required_file.rb
@@ -1,0 +1,1 @@
+DOT_SLASH_REQUIRED_FILE = true